made cli py and moved cli code there from main py
